% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_build.R
\name{binary_boot_impute}
\alias{binary_boot_impute}
\title{Imputation of a baseline binary exposure}
\usage{
binary_boot_impute(
  y_formula,
  e_formula,
  sample_subj,
  full_data,
  strata = c("oc_strata", "dtime"),
  qpoints,
  study_id = "id",
  z_formula,
  exposure = "exposure",
  outcome,
  time = "time",
  sweights = "sampWT",
  n_imp = 10
)
}
\arguments{
\item{sweights}{character scalar - name of the variable holding sample weight data}

\item{y_formula:}{a model formula for the fixed effect portion of the outcomes model (see buildr::logit_mix)}

\item{e_formula:}{a model formula for the marginal exposure  (see the stat::gml function)}

\item{sample_subj:}{subject level dataset, from the sampled data (includes assessed exposure)}

\item{full_data:}{population data that was sampled from}

\item{strata:}{character vector - names of the outcome and dropout variables (in that order)}

\item{qpoints:}{number of quadrature points to integrate the likelihood}

\item{study_id:}{character scalar - cluster identifier}

\item{z_formula:}{a model formula for the random effect portion of the outcomes model (see buildr::logit_mix)}

\item{exposure:}{character scalar - name of the binary exposure variable}

\item{outcome:}{character scalar - name of the outcome variable}

\item{time:}{character - name of the time effect}

\item{n_imp:}{number of imputed datasets to produce}
}
\value{
A list of imputed dataframes
}
\description{
Imputation of a baseline binary exposure. Produces a list of imputed dataframes on the dataset named in the 'full_data' argument.
}
