% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingFunctions.R
\name{build_sample_frame}
\alias{build_sample_frame}
\title{BUILD sample frame}
\usage{
build_sample_frame(data, study_id, outcome, drop_strata)
}
\arguments{
\item{data:}{the name of a dataframe to select rows from}

\item{study_id:}{character, variable in the dataframe that identifies the subject}

\item{outcome:}{character, variable in the dataframe that specifies the study outcome}

\item{drop_strata:}{character - variable in the dataframe that identifies the dropout strata

@return A list of two elements:
\enumerate{
\item sample_frame: a dataframe with one row per sampling unit. columns include the study ID from the original data, the dropout strata membership (drop_strata) and the derived outcome strata (oc_strata).
\item strata_counts: a matrix of unit counts within each outcome (rows) and dropout (columns) strata.
}}
}
\description{
Accepts a longitudinal dataset, computes the outcome strata and outputs a dataframe with the sampling frame for the analysis.
}
\examples{
set.seed(957385)
dat = simuDat(
  B = c(-3, 2, 6, 2, 0),
  G = matrix(c(4), ncol = 1),
  random = NULL,
  p_exposed = 0.5,
  p_confounder = 0.25,
  conf = 0,
  N = 1000,
  M = 5,
  D_B = c(0, -2, -1, -0.5, 0),
  D_F = c(1, 1, 1, 1, 1),
  D_parms = c(1, 1)
) 
mdat <- subset(dat, time <= dtime)

s_frame <- build_sample_frame(data = mdat, study_id = 'id', outcome = 'y', drop_strata = 'dtime')

}
