% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper.R
\name{clustHyper}
\alias{clustHyper}
\title{Write dotplot for hypergeometric test.}
\usage{
clustHyper(dat, clusts, out)
}
\arguments{
\item{dat}{A \code{data.frame} with columns as metrics to be tested for enrichment.}

\item{clusts}{A vector giving the cluster ID of each row in \code{dat}.}

\item{out}{The output file name passed to \code{\link{dotPscale}}.}

\item{...}{Additional arguments to \code{\link{dotPscale}}.}
}
\value{
A list of results from \code{\link{corHeatmap}}.
}
\description{
Write dotplot for hypergeometric test.
}
