% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpset.R
\name{colKeyFn}
\alias{colKeyFn}
\title{Plots color key of UpSet.}
\usage{
colKeyFn(name, colfn)
}
\arguments{
\item{name}{Color key title.}

\item{colfn}{Function for mapping colors.}
}
\description{
Plots color key of UpSet.
}
