% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_build.R
\name{combine_imp}
\alias{combine_imp}
\title{Function to combine results from models on imputed data}
\usage{
combine_imp(mu, vcov)
}
\arguments{
\item{mu:}{a mXp matrix of parameter estimates (p is the number of parameters and m is the number of imputations)}

\item{vcov:}{a pXpXm tensor of covariance matrices}
}
\value{
A two element list:
\enumerate{
\item a vector of the imputation adjusted estimates
\item a variance covariance matrix for the estimates
}
}
\description{
A function to combine results from models on imputed data.
Returns a list, the first element is the mean vector, while element 2
is the variance matrix
}
