% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpset.R
\name{combSel}
\alias{combSel}
\title{Converts intersect names to logical vector.}
\usage{
combSel(id)
}
\arguments{
\item{id}{The intersect name.}
}
\value{
A logical vector
}
\description{
Converts intersect names to logical vector.
}
