% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{cond.col.vec}
\alias{cond.col.vec}
\title{Creates a color scale for the levels in a vector, then returns a vector assigning a color to each element of the input vector.}
\usage{
cond.col.vec(cond, colfn = rainbow, ...)
}
\arguments{
\item{cond}{A vector that can be coerced to a factor.}

\item{colfn}{A function that returns a color map for each level in \code{cond}.}

\item{...}{Additional arguments to \code{colfn}.}
}
\value{
A vector of colors corresponding to the elements of \code{cond}.
}
\description{
Creates a color scale for the levels in a vector, then returns a vector assigning a color to each element of the input vector.
}
