% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_build.R
\name{cond_p_exposure}
\alias{cond_p_exposure}
\title{Compute exposure probabilities conditional on outcomes}
\usage{
cond_p_exposure(
  y_formula,
  z_formula = ~1,
  y_fit,
  e_formula,
  e_fit,
  full_data,
  exposure = "exposure",
  study_id = "id",
  outcome = "y",
  time = "time",
  qpoints
)
}
\arguments{
\item{y_formula:}{a model formula for the fixed effect portion of the outcomes model (see buildr::logit_mix)}

\item{z_formula:}{a model formula for the random effect portion of the outcomes model (see buildr::logit_mix)}

\item{y_fit:}{output object from the outcomes model (see buildr::logit_mix)}

\item{e_formula:}{a model formula for the marginal exposure  (see the stat::gml function)}

\item{e_fit:}{an output object from the marginal exposure model (see the stat::gml function)}

\item{full_data:}{full population data (with undetermined exposure)}

\item{exposure:}{character - name of the exposure variable in the sampled data}

\item{study_id:}{character - name of the cluster id variable in the population and sampled data}

\item{outcome:}{character - name of the study outcome variable in the population and sampled data}

\item{time:}{character - name of the time variable in the population and sampled data}

\item{qpoints:}{integer - number of quadrature points needed to integrate the likelihood}
}
\value{
A vector of probabilities of exposure, in participant order
}
\description{
: A function to compute exposure probabilities conditional on outcomes.
}
