% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcor.R
\name{corHeatmap}
\alias{corHeatmap}
\title{Plots a heatmap of the correlations between all columns in a matrix.}
\usage{
corHeatmap(
  mat,
  filename = "correlation",
  method = "pearson",
  cell.dim = 0.2,
  ...
)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{method}{Method used by \link{\code{cor()}}.}

\item{...}{Additional arguments to \link{\code{Heatmap()}}.}

\item{colfn}{Function to generate color scale. By default uses \link{\code{col.abs()}}.}
}
\value{
A Heatmap object.
}
\description{
Plots a heatmap of the correlations between all columns in a matrix.
}
