% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.apply}
\alias{dir.apply}
\title{Wrapper for applying a write function to each element of list \code{x} and using \code{names(x)} as the filenames.}
\usage{
dir.apply(x, path, fn = dir.tab, ...)
}
\arguments{
\item{x}{A list.}

\item{fn}{The function to be applied to each element of \code{x}. It should accept a \code{path} argument.}

\item{...}{Additional arguments to \code{fn}.}
}
\value{
A list.
}
\description{
Wrapper for applying a write function to each element of list \code{x} and using \code{names(x)} as the filenames.
}
\examples{
x <- split(iris,iris$Species)
dir.apply(x, dir.csv, path="path/to")
}
\seealso{
\code{\link{dir.tab}}, \code{\link{base::lapply}}
}
