% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.csv}
\alias{dir.csv}
\title{Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path as a csv.
Any folders in the path that do not exist are created.}
\usage{
dir.csv(x, filename, path = ".", summary = F, quote = T, ..., append.date = T)
}
\arguments{
\item{x}{Data to be written to \code{filename}.}

\item{filename}{The output file name.}

\item{path}{The path to the output file.}

\item{...}{Additional arguments to \code{write.csv}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\description{
Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path as a csv.
Any folders in the path that do not exist are created.
}
\examples{
dir.csv(iris, "example", path="path/to")
}
\seealso{
\code{\link{dir.out}}, \code{\link{utils::write.csv}}
}
