% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.pdf}
\alias{dir.pdf}
\title{Concatenates a path, filename, and file extension into an output path, then opens a pdf connection at the output path.
Any folders in the path that do not exist are created.}
\usage{
dir.pdf(filename, path = ".", ..., append.date = T)
}
\arguments{
\item{filename}{The output file name.}

\item{path}{The path to the output file.}

\item{...}{Additional arguments to \code{pdf}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\description{
Concatenates a path, filename, and file extension into an output path, then opens a pdf connection at the output path.
Any folders in the path that do not exist are created.
}
\examples{
dir.pdf("example",path="path/to")
plot(1:5,1:5)
dev.off()
}
\seealso{
\code{\link{dir.img}}, \code{\link{grDevices::pdf}}
}
