% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.tab}
\alias{dir.tab}
\title{Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path as a tab-separated file.
Any folders in the path that do not exist are created.}
\usage{
dir.tab(x, filename, path = ".", ext = "txt", quote = F, ..., append.date = T)
}
\arguments{
\item{x}{Data to be written to \code{filename}.}

\item{filename}{The output file name.}

\item{path}{The path to the output file.}

\item{ext}{The file extension.}

\item{...}{Additional arguments to \code{write.table}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\description{
Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path as a tab-separated file.
Any folders in the path that do not exist are created.
}
\examples{
dir.tab(iris, "example", path="path/to")
}
\seealso{
\code{\link{dir.out}}, \code{\link{utils::write.table}}
}
