% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Probability density function for MV normal}
\usage{
dmvnorm(x, mu = NULL, sigma = NULL)
}
\arguments{
\item{x:}{numeric - a vector of observed points for which to compute the pdf value (real numbers)}

\item{mu:}{numeric - a vector of distribution mean values (real numbers)}

\item{sigma:}{numeric - distribution variance matrix (\eqn{\gt 0})}
}
\value{
A scalar representing the height of the normal surface at x
}
\description{
: Probability density function for MV normal for arbitrary mean vector and variance matrix.
If mu and sigma are not specified then a standard MV normal distribution, of the same dimension as x, will be assumed.
}
\examples{

dmvnorm(c(0.25, 0.05), mu = c(0.1, 0.2), sigma = matrix(c(1, 0.25, 0.25, 0.8), 2, 2))
}
