% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_adj_all.R
\name{drop_adj_all}
\alias{drop_adj_all}
\title{Function to apply the dropout correction to multiple parameter estimates}
\usage{
drop_adj_all(
  model_fit,
  aparms = c("(Intercept)", "time", "exposure", "time:exposure"),
  dparms = c("dtime", "time:dtime", "exposure:dtime", "time:exposure:dtime"),
  time = "time",
  exposure = "exposure",
  dtime = "dtime",
  where = c(0, 0, 1, 1)
)
}
\arguments{
\item{model_fit:}{a fit object from logit_mix}

\item{aparms:}{character vector - parameters with hypothesized dropout effects, needing adjustment by dparms}

\item{dparms:}{character vector - dropout parameters to adjust aparms}

\item{time:}{character - study time, with zero indicating baseline}

\item{exposure:}{character - name of the exposure variable, binary (0/1)}

\item{dtime:}{character - name of the variable containing subject dropout time}

\item{where:}{numeric vector of dimension aparms, indicating which subgroup of exposure the dropout time proportion should be calculated}
}
\value{
A list with the following elements: dataframe of model fixed effect estimates, adjusted for dropout effects
\enumerate{
\item \strong{estimate}: a vector of the model fixed effect estimates, adjusted for dropout
\item \strong{vcov}: dropout adjusted Variance-covariance matrix for \strong{estimate}, computed using the delta method
\item \strong{lvec}: the vector used to compute \strong{estimate}
\item \strong{bvec}: derivative vector of \strong{lvec}
\item \strong{bvcov}: block diagonal matrix of the model estimated variance matrix fixed effects and the variance matrix of the dropout time proportions
}
}
\description{
This function is only validated for simple conditional linear models
}
