% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_adjust.R
\name{drop_adjust}
\alias{drop_adjust}
\title{Function to apply the dropout correction to individual parameter estimates}
\usage{
drop_adjust(model_fit, data, var, dvar, where, effect_name, weights = NULL)
}
\arguments{
\item{model_fit:}{a fit object from logit_mix}

\item{data:}{the data used to estimate the model in model_fit}

\item{var:}{character - variable to be adjusted}

\item{dvar:}{character - the dropout specific variable used to adjust}

\item{where:}{Boolean - subset specification for the correct group to calculate
dropout time proportions}

\item{effect_name:}{character - if var is a combination then specify this for the new name}

\item{weights:}{character - name of the sampling weight column in 'data'}
}
\value{
A one row dataframe containing the dropout adjusted fixed effect estimate. Columns include:
\enumerate{
\item \strong{estimate}: estimate of the specified parameter, adjusted for dropout
\item \strong{std_err}: dropout adjusted standard error of \strong{estimate}, computed using the delta method
}
}
\description{
This function is only validated for simple conditional linear models
}
