% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_quadrature.R
\name{gauss_herm_consts}
\alias{gauss_herm_consts}
\title{Points and weight for Gauss-Hermite quadrature}
\usage{
gauss_herm_consts(n, ndim = 1)
}
\arguments{
\item{n:}{number of points to generate}

\item{ndim:}{number of points to generate the dimension of the integration}
}
\value{
A list with numeric matrix components:
\enumerate{
\item \strong{pnts}: quadrature points (ndim by n)
\item \strong{wts}: quadrature weights (1 by n)
}
}
\description{
: This function generates a list of points an weights for Gauss-Hermite quadrature. Since it is anticipated that general use will be to integrate with respect to a normal density, the points are pre-multiplied by the square root of 2 and the weights divided by the square root of pi.
}
\examples{
gauss_herm_consts(5)
}
