% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{gmat}
\alias{gmat}
\title{Accepts a covariance vector and outputs covariance matrix}
\usage{
gmat(cvec, cstruct = "un", chol = F)
}
\arguments{
\item{cvec:}{a vector of covariance parameters}

\item{cstruct:}{character - desired matrix structure ('diag'/'un')}

\item{chol:}{boolean. If true then cvec is considered to be the Cholesky root of the final product matrix.}
}
\value{
A matrix
}
\description{
: This function accepts a covariance vector and outputs covariance matrix.
The assumed structure can be diagonal or unstructured and can compute the matrix assuming that the
parameter vector represents the Cholesky root of the final product.
}
\examples{
a <- c(0.9,1.2)
gmat(a, 'diag', chol = F)
gmat(a, 'diag', chol = T)

b <- c(2, 0.25, 0.5)
gmat(b, 'diag', chol = F)
gmat(b, 'un', chol = F)
gmat(b, 'un', chol = T)
}
