% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_quadrature.R
\name{gquad}
\alias{gquad}
\title{Basic Gaussian Quadrature}
\usage{
gquad(FUNCT, PW, ...)
}
\arguments{
\item{...}{: further arguments to the user supplied function in FUNCT}

\item{FUNCT:}{name of a function to integrate}

\item{PW:}{a list of points and weights}
}
\value{
A scalar result of the integration
}
\description{
: This function implements integration of a function by Gaussian quadrature, for supplied points and weights
}
\examples{

# gauss-hermite quadrature: integrate (1+x^2)*exp((-x^2)/2)/sqrt(2*pi)
f <- function(x) (1+x^2) 
# generate points and weights
ghpw <- gauss_herm_consts(3)
# integrate
gquad(f, ghpw)
# check
f2 <- function(x) (1+x^2)*exp((-x^2)/2)/sqrt(2*pi)
integrate(f2, -Inf, Inf)
}
