% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_quadrature.R
\name{gquad_cntr_scale}
\alias{gquad_cntr_scale}
\title{Center and scale quadrature points}
\usage{
gquad_cntr_scale(pw, mu, vcov)
}
\arguments{
\item{pw:}{a point and weight object from gauss_herm_consts()}

\item{mu:}{a mean vector used to center the points}

\item{vcov:}{a variance matrix used to scale the points}
}
\value{
A centered and scaled version of the input from gauss_herm_consts()
}
\description{
: This function accepts an object from  gauss_herm_consts() and scales the points for a supplied mean vector and variance matrix
}
\examples{
ghpw <- gauss_herm_consts(3, 2)
mu <- c(1, 2)
vcov <- matrix(c(2, 0.25, 0.25, 0.5), 2, 2)
ghpw_new <- gquad_cntr_scale(ghpw, mu, vcov)
}
