% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_quadrature.R
\name{herm_coef}
\alias{herm_coef}
\title{coefficients of the Hermite polynomial (Salzer 1952)}
\usage{
herm_coef(n)
}
\arguments{
\item{n:}{number of points to generate}
}
\value{
a numeric vector
}
\description{
: This function generates coefficients of the Hermite polynomial (Salzer 1952)
Salzer H. E., Zucker R., Capuano R. "Table of the Zero and Weight Factors of the First Twenty Hermite Polynomials" Journal of Research of the National Bureau of Standards, Vol 48, No. 2, February 1952
}
\examples{
herm_coef(5)
}
