% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_quadrature.R
\name{herm_wt}
\alias{herm_wt}
\title{Weights for the Gauss-Hermite quadrature points (Salzer 1952)}
\usage{
herm_wt(x, n)
}
\arguments{
\item{x:}{coefficients of the Hermite polynomial}

\item{n:}{Number of coefficients of the Hermite polynomial}
}
\value{
a numeric vector
}
\description{
: This function generates weights for the Gauss-Hermite quadrature points (Salzer 1952).
Salzer H. E., Zucker R., Capuano R. "Table of the Zero and Weight Factors of the First Twenty Hermite Polynomials" Journal of Research of the National Bureau of Standards, Vol 48, No. 2, February 1952
}
\examples{
hx <- Re(polyroot(herm_coef(4)))
sapply(hx, herm_wt, n = 4)
}
