% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{hm.ann}
\alias{hm.ann}
\title{Wrapper for \code{hm.cell()} which applies cell dimensions to simple \code{rowAnnotation}s}
\usage{
hm.ann(x, ..., conds = NULL, heatmap_width = NULL, cell.w = NULL)
}
\arguments{
\item{x}{A numeric matrix to be plotted.}

\item{...}{Additional arguments to \code{hm.cell()}.}

\item{conds}{A matrix of factors to be used as \code{rowAnnotation}s.}

\item{heatmap_width}{The heatmap width Ignored if \code{cell.w} is specified.}

\item{cell.w}{The cell width.}
}
\value{
A ComplexHeatmap.
}
\description{
Wrapper for \code{hm.cell()} which applies cell dimensions to simple \code{rowAnnotation}s
}
