% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdot.R
\name{hyperDot}
\alias{hyperDot}
\title{accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results}
\usage{
hyperDot(odds, fdr, q, ...)
}
\arguments{
\item{odds}{A matrix of log2 odds ratios from an enrichment test.}

\item{fdr}{A matrix of FDR values from an enrichment test.}

\item{q}{A matrix of test counts for each test.}

\item{...}{Additional arguments to \code{hmdot()}.}
}
\description{
accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results
}
