% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_build.R
\name{imputeit}
\alias{imputeit}
\title{A convenience function to run the imputation, analyze the data and combine results}
\usage{
imputeit(
  y_formula,
  e_formula,
  sample_subj,
  full_data,
  strata,
  study_id,
  sweights,
  z_formula,
  exposure,
  outcome,
  time,
  n_imp = 30,
  qpoints = 10,
  aparms = c("(Intercept)", "time", "exposure", "time:exposure"),
  dparms = c("dtime", "time:dtime", "exposure:dtime", "time:exposure:dtime"),
  dtime = "dtime",
  where = c(0, 0, 1, 1)
)
}
\arguments{
\item{y_formula:}{a model formula for the fixed effect portion of the outcomes model (see buildr::logit_mix)}

\item{e_formula:}{a model formula for the marginal exposure  (see the stat::gml function)}

\item{sample_subj:}{subject level dataset, from the sampled data (includes assessed exposure)}

\item{full_data:}{population data that was sampled from}

\item{strata:}{character vector - names of the outcome and dropout variables (in that order)}

\item{qpoints:}{number of quadrature points to integrate the likelihood}

\item{study_id:}{character scalar - cluster identifier}

\item{z_formula:}{a model formula for the random effect portion of the outcomes model (see buildr::logit_mix)}

\item{exposure:}{character scalar - name of the binary exposure variable}

\item{outcome:}{character scalar - name of the outcome variable}

\item{time:}{character - name of the time effect (scaled such that 0 )}

\item{sweights:}{character scalar - name of the variable holding sample weight data}

\item{n_imp:}{number of imputed datasets to produce}
}
\value{
A list. The first element contains a vector of dropout corrected regression coefficients. Element 2 is a matrix of regression coefficients, each row from a separate imputation.
The third element is the corrected variance matrix for the coefficients in element 1. Element 4 is an array with three dimensions containing the variance matrix from each imputation. The third dimension indexes the imputation.
The final element is a list of the fit object from each imputation.
}
\description{
A convenience function to run the imputation, analyze the data and combine results
}
\examples{
set.seed(957385)
dat = simuDat(
  B = c(-3, 2, 6, 2, 0),
  G = matrix(c(4), ncol = 1),
  random = NULL,
  p_exposed = 0.5,
  p_confounder = 0.25,
  conf = 0,
  N = 1000,
  M = 5,
  D_B = c(0, -2, -1, -0.5, 0),
  D_F = c(1, 1, 1, 1, 1),
  D_parms = c(1, 1)
) 
mdat <- subset(dat, time <= dtime)

s_frame <- build_sample_frame(data = mdat, study_id = 'id', outcome = 'y', drop_strata = 'dtime')

s_plan <- build_sample_plan(n_samp = c(50, 200, 50), N_bystrata = s_frame$strata_counts, drop_grad = 1)

s_sample <- strat_samp(s_frame$sample_frame,oc_strata = 'oc_strata', drop_strata = 'dtime', sample_plan=s_plan$sample_plan,sample_by = 'n')

# get exposure status
 sample <- merge(mdat[mdat$time ==0, ]
              , s_sample[, c('id'
                         , 'oc_strata'
                         , 'drop_strata'
                         , 'sampWT'
                        )]
              , by = 'id')
# MI analysis
 im_fit <- imputeit(y_formula = y ~ time + exposure + time:exposure 
                  , e_formula = exposure ~ 1
                  ,  strata='oc_strata'
                  , sample_subj  = sample
                  , full_data = mdat 
                  , study_id = 'id', sweights='sampWT', z_formula = ~1, exposure = 'exposure'
                  , outcome='y', time = 'time' 
                  , n_imp = 5
                  , qpoints = 20)
}
