% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_build.R
\name{L_i_logit}
\alias{L_i_logit}
\title{likelihood for a logit model, Independence}
\usage{
L_i_logit(Yi, Pi)
}
\arguments{
\item{Yi:}{binary outcome (1/0) for the ith observation}

\item{Pi:}{probability of the outcome}
}
\value{
A scalar
}
\description{
: Returns the likelihood for a single observation.
}
\examples{
L_i_logit(1, 0.25)
L_i_logit(0, 0.25)
}
