% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{L_i_logit_glmm}
\alias{L_i_logit_glmm}
\title{Compute the likelihood for a binary logistic model for a single cluster.}
\usage{
L_i_logit_glmm(Yi, Pi, ghqpw)
}
\arguments{
\item{Yi:}{a vector outcomes for the cluster}

\item{Pi:}{a vector probabilities for each measurement}

\item{ghqpw:}{a list of points and weights to integrate the likelihood}
}
\value{
scalar
}
\description{
: Compute the likelihood for a binary logistic model for a single cluster.
}
\examples{
L_i_logit_glmm(Yi, Pi, ghqpw)
}
