% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper.R
\name{lHyper}
\alias{lHyper}
\title{Hypergeometric test for enrichment of conditions in a cluster.}
\usage{
lHyper(bg, test, clusts, padj.method = "fdr")
}
\arguments{
\item{bg}{A vector of sample IDs.}

\item{test}{A list of vectors of IDs to be tested}

\item{clusts}{A list of vectors of IDs}

\item{padj.method}{Method passed to \code{\link{p.adjust}} for multiple hypothesis correction.}
}
\description{
Hypergeometric test for enrichment of conditions in a cluster.
}
\seealso{
\code{\link{phyper2}}, \code{\link{p.adjust}}
}
