% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{ll_i_logitGlmm}
\alias{ll_i_logitGlmm}
\title{Compute the log-likelihood for a binary logistic model for a single cluster.}
\usage{
ll_i_logitGlmm(
  parms,
  Xi,
  Yi,
  Zi,
  cstruct,
  ghqpw,
  acml = F,
  SP = NULL,
  OS = NULL,
  DS = NULL
)
}
\arguments{
\item{parms:}{a vector of parameters, fixed effects followed by covariance parameters}

\item{Xi:}{a matrix of observed fixed data for the cluster}

\item{Yi:}{a vector outcomes for the cluster}

\item{Zi:}{a matrix of random effects for the cluster}

\item{cstruct:}{covariance structure for the random effects}

\item{ghqpw:}{a list of points and weights to integrate the likelihood}

\item{acml:}{boolean - if T then use ascertainment correction}

\item{SP:}{matrix of sampling probabilities 3XD - D is the number of dropout categories}

\item{OS:}{Outcome strata of the cluster (1-3)}

\item{DS:}{integer dropout strata of the cluster (1-D).}
}
\value{
scalar
}
\description{
: Compute the log-likelihood for a binary logistic model for a single cluster. Option to add ascertainment correction.
}
\examples{
ll_i_logitGlmm(parms, Xi, Yi, Zi, cstruct, ghqpw, acml = F, SP, OS, DS)
}
