% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{logit_mix}
\alias{logit_mix}
\title{GLMM for binary outcomes, using cluster sample weighting and ascertainment correction}
\usage{
logit_mix(
  df,
  fixed,
  random = ~1,
  cluster,
  sweights = NULL,
  acml = F,
  sprobs = NULL,
  outcome_strata = NULL,
  drop_strata = NULL,
  cstruct = "un",
  robust = F,
  repred = F,
  qpoints = 10,
  starting = NULL,
  opt_method = "BFGS",
  opt_control = list(),
  adapt = F
)
}
\arguments{
\item{df:}{a data frame}

\item{fixed:}{formula for the fixed effect portion of the model}

\item{random:}{formula for the fixed effect portion of the model}

\item{cluster:}{character - name of the variable in df that defines a cluster}

\item{sweights:}{character - name of the variable in df that holds sampling weights (cluster level)}

\item{acml:}{boolean - if true than use ascertainment corrected likelihood}

\item{sprobs:}{a matrix of sampling probabilities dimension 3XD - D being the number of dropout categories}

\item{outcome_strata:}{character - name of the variable in df that contains the cluster level outcome strata}

\item{drop_strata:}{character - name of the variable in df that contains the cluster level dropout strata}

\item{cstruct:}{text (diag, un) - covariance structure for the random effects}

\item{robust:}{boolean - if true then compute robust variances for the fixed effects}

\item{rpred:}{boolean - if true add a prediction of the random cluster random effects to the output object}

\item{qpoints:}{numeric scalar - number of quadrature points to use when integrating the likelihood}

\item{starting:}{numeric vector - starting values fixed followed by random. choice of cstruct and random arguments determine form}

\item{opt_method:}{character - optimization method (see optim function)}

\item{opt_control:}{list - arguments to fine tune the optimization (see optim function)}

\item{adapt:}{boolean - if true then use adaptive quadrature - this is in development, slow and does not converge well}
}
\value{
A list of regression products
func_call: the function call that created the model object
data: a copy of the dataframe used in the analysis, sorted the cluster ID
opt_fit: the output from the optim function for the model optimization
vcov: the variance covariance matrix of the fixed effects and parameters of the variance components (Cholesky decomposition)
G: the estimated covariance matrix of the random effects
rvcov: (if robust = T) the robust covariance matrix of the fixed effects
fixed: a data frame containing the estimates (estimate) standard error (std_err) and test statistic (wald_stat) of the fixed effect parameter estimates
}
\description{
: GLMM for binary outcomes, using cluster sample weighting and ascertainment correction. Robust variance estimates of the fixed effects are available. Derivatives of the log-likelihood are computed using the numDeriv package.
}
\examples{
set.seed(957385)
dat = simuDat(
  B = c(-3, 2, 6, 2, 0),
  G = matrix(c(4), ncol = 1),
  random = NULL,
  p_exposed = 0.5,
  p_confounder = 0.25,
  conf = 0,
  N = 1000,
  M = 5,
  D_B = c(0, -2, -1, -0.5, 0),
  D_F = c(1, 1, 1, 1, 1),
  D_parms = c(1, 1)
) 
mdat <- subset(dat, time <= dtime)

l_mix_fit <- logit_mix(df=mdat, fixed=y~time, random = ~ 1, cluster='id')
}
\references{
Gilbert, P. and Varadhan, R. (2025). numDeriv: Accurate Numerical Derivatives. R package version 2022.9-1.
}
