% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{lrtab}
\alias{lrtab}
\title{Applies a function to all file names in a specified directory matching a given pattern.}
\usage{
lrtab(dir, fn = read.table, pattern = NULL, ...)
}
\arguments{
\item{dir}{The traget directory.}

\item{fn}{The function to be applied to each directory name.}

\item{pattern}{The regular expression matching desired files.}

\item{...}{Additional arguments to \code{fn}.}
}
\value{
A list.
}
\description{
Applies a function to all file names in a specified directory matching a given pattern.
}
\examples{
dir.csv(iris, "example", path="path/to")
dir.tab(iris, "example", path="path/to")
lrtab("path/to", read.csv, "\\\\.csv")
}
\seealso{
\code{\link{dir.img}}, \code{\link{base::list.files}}
}
