% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mat_spec_sqrt}
\alias{mat_spec_sqrt}
\title{Square root of a matrix, by spectral decomposition}
\usage{
mat_spec_sqrt(A)
}
\arguments{
\item{A:}{a matrix}
}
\value{
A matrix
}
\description{
: Accepts a matrix as input and returns the square root using spectral decomposition
}
\examples{
vcov <- matrix(c(2, 0.25, 0.25, 0.5), 2, 2)
mat_spec_sqrt(vcov)

}
