% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{n2ll_logit_glmm}
\alias{n2ll_logit_glmm}
\title{Compute the -2log-likelihood for a binary logistic model for a full dataset.}
\usage{
n2ll_logit_glmm(
  parms,
  X,
  Y,
  Z,
  SW,
  cstruct,
  ghqpw,
  acml = F,
  SP = NULL,
  OS = NULL,
  DS = NULL,
  ll_only = T,
  adapt = F
)
}
\arguments{
\item{parms:}{a vector of parameters, fixed effects followed by covariance parameters}

\item{X:}{list of matrices of observed fixed data for the cluster}

\item{Y:}{a list of a vectors of outcomes}

\item{Z:}{a list of matrices of random effects}

\item{cstruct:}{covariance structure for the random effects}

\item{ghqpw:}{a list of points and weights to integrate the likelihood}

\item{acml:}{boolean - if T then use ascertainment correction}

\item{SP:}{matrix of sampling probabilities 3XD - D is the number of dropout categories}

\item{OS:}{a vector of outcome strata of the cluster (1-3)}

\item{DS:}{a vector of integer dropout strata of the cluster (1-D).}
}
\value{
scalar
}
\description{
: Compute the -2log-likelihood for a binary logistic model for a full dataset.
}
\examples{
n2ll_logit_glmm(parms, X, Y, Z, SW, cstruct, ghqpw, acml = F
, SP = NULL # matrix of sampling probabilities 3XD
, OS = NULL # Outcome strata
, DS = NULL # dropout strata
, ll_only = T
, adapt = F
)
}
