% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_mix.R
\name{n2ll_rparm_i_logitGlmm}
\alias{n2ll_rparm_i_logitGlmm}
\title{Cluster -2*log-likelihood contribution for specified random parameters.}
\usage{
n2ll_rparm_i_logitGlmm(re_i, G, XB, Yi, Zi)
}
\arguments{
\item{re_i:}{random parameters for cluster i}

\item{G:}{covariance parameters of the random effects}

\item{XB:}{vector of fixed effect variables multiplied by corresponding vector of parameters}

\item{Yi:}{a vector outcomes for the cluster}

\item{Zi:}{a matrix of random effects for the cluster}
}
\value{
scalar
}
\description{
: Cluster -2*log-likelihood contribution for specified random parameters.
}
\examples{
n2ll_rparm_i_logitGlmm(re_i, G, XB, Yi, Zi)
}
