% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{np_boot}
\alias{np_boot}
\title{Single non-parametric bootstrap}
\usage{
np_boot(df, ...)
}
\arguments{
\item{df:}{a dataframe to select a bootstrap sample from}
}
\value{
A bootstrap resampled version of the submitted dataframe with an added psuedo-identifier (.rsUnit.)
}
\description{
Returns a single iteration non-parametric bootstrap data frame. The new data frame will include an integer record key: .rsUnit.
}
\examples{

bwdat <- MASS::birthwt

bdat <- np_boot(bwdat)
}
