% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{np_clus_boot}
\alias{np_clus_boot}
\title{Single non-parametric cluster bootstrap}
\usage{
np_clus_boot(df, cluster, ...)
}
\arguments{
\item{df:}{a dataframe to select a bootstrap sample from}

\item{cluster:}{character, name of a variable that defines cluster membership}
}
\value{
Output:
A cluster bootstrap version of the input data, with an additional variable meant as
a synthetic cluster ID that can be used validly as input to a function that uses
clustering as part of the algorithm: .ClusID.
}
\description{
returns a single iteration non-parametric cluster bootstrap data frame
}
\examples{
set.seed(957385)
dat = simuDat(
  B = c(-3, 2, 6, 2, 0),
  G = matrix(c(4), ncol = 1),
  random = NULL,
  p_exposed = 0.5,
  p_confounder = 0.25,
  conf = 0,
  N = 1000,
  M = 5,
  D_B = c(0, -2, -1, -0.5, 0),
  D_F = c(1, 1, 1, 1, 1),
  D_parms = c(1, 1)
) 
mdat <- subset(dat, time <= dtime)

cdat <- np_clus_boot(mdat, 'id')
}
