% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{pbetabinom}
\alias{pbetabinom}
\title{Probability density function for a beta-binomial distribution}
\usage{
pbetabinom(n, x, a, b)
}
\arguments{
\item{n:}{numeric scalar - the number of trials (\eqn{\ge 0})}

\item{x:}{numeric scalar - the number of successes (\eqn{\ge 0})}

\item{a:}{numeric scalar - shape parameter (\eqn{\gt 0})}

\item{b:}{numeric scalar - shape parameter (\eqn{\gt 0})}
}
\value{
A scalar representing the PDF value at chosen parameters
}
\description{
: Probability density function for a beta-binomial distribution given by:

\eqn{{n \choose k} \frac{\Gamma(a+b)\Gamma(x+a)\Gamma(n-x+b)}{\Gamma(a)\Gamma(b)\Gamma(a+b+n)}}

Where \eqn{\Gamma()} is the gamma function.

Reference: Casella G., Berger R. L. "Statistical Inference, Second Edition" Duxbury Press, 2002
}
\examples{

sapply(0:4, function(x){pbetabinom(4,x,1.5,1)})

}
