% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingFunctions.R
\name{samp_fun}
\alias{samp_fun}
\title{Sample selection}
\usage{
samp_fun(df, p = NULL, n = NULL)
}
\arguments{
\item{df:}{the name of a dataframe to select rows from}

\item{p:}{numeric (0-1) selection probability}

\item{n:}{integer number to be randomly selected}
}
\value{
A dataframe consisting of a random sample of rows from the original.
}
\description{
Sampling from a data frame without replacement
}
\examples{

bwdat <- MASS::birthwt

sample_dat <- samp_fun(bwdat, n = 20)

}
