% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingFunctions.R
\name{strat_samp}
\alias{strat_samp}
\title{Stratified sample selection}
\usage{
strat_samp(
  df,
  oc_strata = NULL,
  drop_strata = NULL,
  sample_plan,
  sample_by = "n"
)
}
\arguments{
\item{df:}{the name of a dataframe to select rows from}

\item{oc_strata:}{character, variable in the dataframe that specifies the outcome strata membership}

\item{drop_strata:}{character, variable in the dataframe that specifies the dropout strata membership}

\item{sample_plan:}{dataframe - contains the sampling plan from 'build_sample_plan'}

\item{sample_by:}{(p/n) specify whether to sample by counts or probability (default is 'n')}
}
\value{
A dataframe consisting of a stratified random sample of rows from the original.
}
\description{
Stratified sampling from a data frame without replacement.
Within each stratum a fixed proportion of rows will be drawn
}
\examples{

bwdat <- MASS::birthwt

table(bwdat$race)

s_plan <- data.frame(race=c(1,2,3), n = c(20, 25, 15), n_population = as.numeric(table(bwdat$race)))
s_plan <- within(s_plan, {p <- n/n_population})

sample_dat <- strat_samp(bwdat, oc_strata = 'race', sample_plan = s_plan,sample_by='n')

}
