% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{strata_boot}
\alias{strata_boot}
\title{Single stratified non-parametric bootstrap}
\usage{
strata_boot(df, strata, ...)
}
\arguments{
\item{df:}{a dataframe to select a bootstrap sample from}

\item{strata:}{character, name of a variable that defines stratum  membership}
}
\value{
A bootstrap resampled (within specified strata) version of the submitted dataframe with an added psuedo-identifier (.rsUnit.)
}
\description{
Returns a stratified non-parametric bootstrap data frame.
}
\examples{

bwdat <- MASS::birthwt

sbdat <- strata_boot(bwdat, 'smoke')

table(bwdat$smoke)
table(sbdat$smoke)
}
