% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHeatmap.R
\name{writeHeatmap}
\alias{writeHeatmap}
\title{Wrapper for \code{Heatmap()} which allows automatic scaling to specified cell dimensions and character width, and writes the output to an appropriately sized imge file.}
\usage{
writeHeatmap(
  mat,
  filename,
  ...,
  fn = Heatmap,
  conds = NULL,
  left_annotation = NULL,
  cell.h = NULL,
  cell.w = NULL,
  char.h = NULL,
  char.w = NULL,
  height = NULL,
  width = NULL,
  heatmap_height = NULL,
  heatmap_width = NULL,
  units = "in",
  path = ".",
  append.date = F,
  imgfn = dir.pdf,
  buffer.h = 0,
  buffer.w = 0
)
}
\arguments{
\item{mat}{A numeric matrix to be plotted.}

\item{filename}{The output file name.}

\item{...}{Additional arguments to plotting function.}

\item{fn}{The plotting function. It should extend \code{\link{ComplexHeatmap::Heatmap()}}}

\item{conds}{A matrix of factors to be used as \code{rowAnnotation}s. Replaces \code{left_annotation}.}

\item{cell.h}{The cell height.}

\item{cell.w}{The cell width.}

\item{char.h}{The characther width for column names.}

\item{char.w}{The character width for row names.}

\item{height}{The heatmap height. Ignored if \code{cell.h} is specified.}

\item{width}{The heatmap width. Ignored if \code{cell.w} is specified.}

\item{heatmap_height}{The heatmap height. Ignored if \code{cell.h} is specified.}

\item{heatmap_width}{The heatmap width. Ignored if \code{cell.w} is specified.}

\item{units}{The unit scale to be used for \code{cell.h} and \code{cell.w}.}

\item{path}{The output directory.}

\item{append.date}{Whether to write the output to a subdirectory with today's date.}

\item{imgfn}{The function to write the image. It should extend \link{\code{dirfns::dir.out()}}}

\item{buffer.h}{Buffer height.}

\item{buffer.w}{Buffer width.}
}
\value{
A ComplexHeatmap.
}
\description{
Wrapper for \code{Heatmap()} which allows automatic scaling to specified cell dimensions and character width, and writes the output to an appropriately sized imge file.
}
